Try Catch With transaction
Try Catch With transaction tanto en citas como en casos/casos status

+++ database/migrations/2022_05_09_093848_create_case_types_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
 
             $table->string('name');
-            $table->bigInteger('amount')->default('100');
+            $table->float('amount')->default('100');
             $table->string('case_category');
             $table->enum('is_active',array('Yes','No'))->default('Yes');
